refactor(celery): consolidate task queues to a single default queue with priority support
