+++ backend/app/celery/utils.py
@@ -218,7 +218,7 @@ def start_tts_result_processing(
 
 def start_fast_evaluation(eval_run_id: int, trace_id: str = "N/A") -> str:
     """Enqueue the run_evaluation_fast orchestrator task for one EvaluationRun."""
-    from app.celery.tasks.evaluation_fast import run_evaluation_fast
+    from app.celery.tasks.job_execution import run_evaluation_fast
 
     task_id = _enqueue_with_trace_context(
         run_evaluation_fast,
